server/requirements.txt
numpy
pandas
tensorflow
Pillow
joblib
scikit-learn
flask
flask-cors
python-dotenv
pefile
Flask-SQLAlchemy